tools/iam-role-mapper-irm/requirements.txt
boto3
collections
pandas
python-docx
pyyaml
google-api-python-client